requirements.txt
streamlit
torch
torchvision
timm
shap
pillow
numpy
matplotlib